+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2.0.0] - 2022-01-31
+### Added
+- Closing connections when event `janiscommerce.ended` was emitted
+- Events listened with `@janiscommerce/events`
+- Using `mongodb@4.3.1` drive
 
 ## [1.18.0] - 2021-07-07
 ### Added